backend/requirements.txt
flask
requests
flask-cors
python-dotenv
pandas
